+++ sqlstore/margin_levels.go
@@ -14,22 +14,19 @@ type AccountSource interface {
 
 type MarginLevels struct {
 	*ConnectionSource
-	columns      []string
-	marginLevels []*entities.MarginLevels
-	batcher      MapBatcher[entities.MarginLevelsKey, entities.MarginLevels]
+	columns       []string
+	marginLevels  []*entities.MarginLevels
+	batcher       MapBatcher[entities.MarginLevelsKey, entities.MarginLevels]
 	accountSource AccountSource
-
 }
 
 const (
-	sqlMarginLevelColumns = `market_id,asset_id,party_id,timestamp,maintenance_margin,search_level,initial_margin,collateral_release_level,vega_time`
+	sqlMarginLevelColumns = `account_id,timestamp,maintenance_margin,search_level,initial_margin,collateral_release_level,vega_time`
 )
 
-func NewMarginLevels(connectionSource *ConnectionSource,
-	accountSource AccountSource) *MarginLevels {
+func NewMarginLevels(connectionSource *ConnectionSource) *MarginLevels {
 	return &MarginLevels{
 		ConnectionSource: connectionSource,
-		accountSource: accountSource,
 		batcher: NewMapBatcher[entities.MarginLevelsKey, entities.MarginLevels](
 			"margin_levels",
 			entities.MarginLevelsColumns),
@@ -43,17 +40,6 @@ func (ml *MarginLevels) Add(marginLevel entities.MarginLevels) error {
 
 func (ml *MarginLevels) Flush(ctx context.Context) error {
 	return ml.batcher.Flush(ctx, ml.pool)
-}
-
-func (t *Transfers) GetTransfersFromParty(ctx context.Context, partyId entities.PartyID) ([]*entities.Transfer, error) {
-	transfers, err := t.getTransfers(ctx,
-		"where transfers_current.from_account_id  in (select id from accounts where accounts.party_id=$1)", partyId)
-
-	if err != nil {
-		return nil, fmt.Errorf("getting transfers from party:%w", err)
-	}
-
-	return transfers, nil
 }
 
 func (ml *MarginLevels) GetMarginLevelsByID(ctx context.Context, partyID, marketID string, pagination entities.Pagination) ([]entities.MarginLevels, error) {